refactor(components): Rename Panels to Panel and simplify structure
This commit refactors the `Panels` component to align with `shadcn/ui` naming conventions and simplifies its structure by using the existing `Card` component.

Key changes include:
- Renamed `Panels` component to `Panel`.
- Renamed `panels` directories and files to `panel`.
- Removed unnecessary sub-components (`PanelHeader`, `PanelTitle`, `PanelDescription`, `PanelSection`, `PanelSectionTitle`).
- Replaced custom Panel sub-components with existing `shadcn/ui` `Card` components.
- Updated all affected imports and JSX usages across the codebase.

This is a breaking change that required updating all instances where the component was used.

Closes #214
