+++ src/components/record/panel/substitutes/index.tsx
